pyproject.toml
[tool.poetry]
name = "gpt4all-voice"
version = "0.1.0"
description = ""
authors = ["Joe DiPilato"]
readme = "README.md"
packages = [{include = "gpt4all_voice"}]

[tool.poetry.dependencies]
python = "~3.10"
sounddevice = "^0.4.6"
ffmpeg-python = "^0.2.0"
whisper-openai = "^1.0.0"
argparse = "^1.4.0"
structlog = "^22.3.0"
pydantic = "^1.10.7"
torchaudio = "^2.0.1"
nomic = "^1.1.4"
# peft = { url = "https://github.com/nomic-ai/nomic/raw/main/bin/peft-0.3.0.dev0-py3-none-any.whl" }
# transformers = { url = "https://github.com/nomic-ai/nomic/raw/main/bin/transformers-4.28.0.dev0-py3-none-any.whl" }

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.2.2"
pydocstyle = "^6.3.0"
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
